backend/requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache
redis==5.0.3
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Validation
pydantic==2.7.0
pydantic-settings==2.2.1
email-validator==2.1.1

# AI / ML
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
scipy==1.13.0
implicit==0.7.2

# HTTP Client
httpx==0.27.0

# Logging
loguru==0.7.2
structlog==24.1.0

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
factory-boy==3.3.0

# Utilities
python-slugify==8.0.4
Pillow==10.3.0
boto3==1.34.82
